% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love.R
\name{degrade}
\alias{degrade}
\title{DE-grade}
\usage{
degrade(g)
}
\arguments{
\item{g}{grade from 0 to 100}
}
\value{
The corresponding 'German grade'
}
\description{
German grades start at 1 (the best) and descend to 5 in uneven increments:
1, 1.3, 1.7, 2, 2.3, 2.7, 3, 3.3, 3.7, 4, 5.
}
\details{
This function assigns german grades to percentage grades.
}
\examples{
  table(degrade(0:100))

}
