% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love.R
\name{degradechart}
\alias{degradechart}
\title{Show a Percentage vs DE Grade Chart}
\usage{
degradechart()
}
\value{
a data.frame of percentage grades from 50 and their
corresponding DE grade
}
\description{
Show a Percentage vs DE Grade Chart
}
