% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hertielove-package.R
\docType{package}
\name{hertielove-package}
\alias{hertielove-package}
\title{#HertieLove}
\description{
Spreads the #HertieLove over your grade and converts percentages to German grades.
}
\author{
\strong{Maintainer}: Will Lowe \email{lowe@hertie-school.org}

}
\keyword{internal}
