% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love.R
\name{hertielove}
\alias{hertielove}
\title{Apply Hertie Love to a Grade Distribution}
\usage{
hertielove(r, plot = FALSE)
}
\arguments{
\item{r}{A vector of grades (0 to 100)}

\item{plot}{whether to plot the new grades against pre-loved grades}
}
\value{
a vector of new grades (0 to 100), returned invisibly
}
\description{
This function figures out which 2PL IRT model would give the
required Hertie grade distribution, assuming if all questions
share the same difficulty parameter (constrained to be positive).
}
\details{
Hertie requires that (no more than) five percent of the class
score below 70, (not more than) five percent of the class score
above 95, and the class median grade is 85.

The function reports SSE: how close it got to fitting those constraints,
alpha and beta, the back estimated IRT parameters, and a comparison table
of the original grade statistics, the required statistics, and those
statistics in the new grades.

The alpha parameter is probably of most interest as it represents the
general level of question difficulty. Values far from zero indicate that
questions are much too easy, or more likely, much too hard (relative to
probable actual abilities).
}
\examples{
  gr <- rbinom(25, prob = 0.66, size = 100)
  newgr <- hertielove(gr)

}
