% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love.R
\name{hertielove}
\alias{hertielove}
\title{Apply Hertie Love to a Grade Distribution}
\usage{
hertielove(r, plot = FALSE)
}
\arguments{
\item{r}{A vector of grades, each from 1 to 100}

\item{plot}{whether to plot the new grades against originals grades}
}
\value{
a vector of new grades, each from 1 to 100, returned invisibly
}
\description{
Students have raw grades in vector r on a 100 point scale.
Rescale this to 0,1 and treat this new score p as the proportion of
questions they answered correctly. The empirical distribution of p
has the 0.05, 0.5, and 0.95 quantiles; label these q_1, q2, q3.
Hertie suggests that q1 should be at 70, q2 should be at 85, and
q3 should be at 95. Probably they are not.
}
\details{
Estimate student ability on a log-odds scale as l = log(p/(1-p)).
If these estimates were from a 2PL IRT model in which the log odds of
answering correctly was (a - d) b where a is ability, d difficulty,
and b the specificity of every question.  Normally we would take the
student scores, estimate the d and b parameters and infer a. To
curve the student scores we will set a = l and choose d and b parameters
to have the preferred quantiles noted above (the optimisation treats
the sum of squared differences between the quantiles implied by a
particular setting of d and b, and q1, q2, and q3.

If plot is TRUE then various before and after comparisons are plotted.
The second bar graph plots the implications for German grades of the
curving. To apply a German grade transformation to any 1-100 grade
distribution, use the \code{degrade} function.
}
\examples{
  gr <- rbinom(25, prob = 0.66, size = 100)
  newgr <- hertielove(gr)

}
